Synched Thread
Syched the thread so that the numbers appear in order. Testing the first
100 numbers seems to have gone much faster although I did not do any
testing and that is just based on my observation. 

Counter seems to work and now it is a matter of letting the program run.

TODO: research how many cores can be running effciently at the same time.
Signed-off-by: bunkerrh <ryanhbunker@gmail.com>